Move two text fields to sign up confirmation
